modularized app targets, abstracted targets visibility for chps and supervisors
