+++ translations/messages-en.properties
@@ -1,4 +1,4 @@
-target.follow_up_household_member = Total Followups
+target.follow_up_household_member = Total Follow Ups
 target.follow_up_household_member.subtitle = This Month
 target.register_households = Total Registered Households 
 target.register_households.subtitle = All Time
